Simplify setup
