--- .babelrc
@@ -1,7 +1,0 @@
-{
-  "sourceMaps": "inline",
-  "comments": false,
-  "presets": [
-    "flow-node"
-  ]
-}
