--- .flowconfig
@@ -1,14 +1,0 @@
-[ignore]
-.*/node_modules/babel.*
-.*/node_modules/tap/*
-.*/node_modules/documentation/*
-.*/node_modules/kefir/*
-
-[libs]
-
-[include]
-
-[options]
-module.name_mapper='unreachable' -> '<PROJECT_ROOT>/src/unreachable'
-suppress_comment= \\(.\\|\n\\)*\\@FlowIssue
-suppress_comment= \\(.\\|\n\\)*\\@FlowIgnore
