+++ .gitignore
@@ -1,8 +1,1 @@
-lib
-dist
-**/*.js
-**/*.flow
-!src/**/*.js
-!src/**/*.flow
-!test/**/*.js
-!test/**/*.flow
+/node_modules/
