--- .npmignore
@@ -1,6 +1,0 @@
-*~
-~*
-!dist
-!lib
-!**/*.js
-!**/*.flow
