+++ .travis.yml
@@ -1,7 +1,9 @@
 language: node_js
-sudo: false
 node_js:
-- 'stable'
+  - '8.10.0'
+  - '8'
+  - '10'
+  - '12'
 deploy:
   provider: npm
   email: rfobic@gmail.com
