+++ index.d.ts
@@ -1,0 +1,2 @@
+declare function unreachable (value: never): never
+export = unreachable
