+++ package.json
@@ -1,6 +1,6 @@
 {
-  "version": "0.2.0",
   "name": "unreachable",
+  "version": "0.2.0",
   "description": "Utility function for exhaustiveness checking with typed JS (TS or Flow)",
   "keywords": [
     "unreachable",
@@ -16,39 +16,18 @@
     "switch",
     "default"
   ],
-  "main": "lib/unreachable",
-  "module": "src/unreachable",
-  "types": "src/unreachable",
-  "author": "Irakli Gozalishvili <rfobic@gmail.com> (http://jeditoolkit.com)",
-  "repository": "https://github.com/Gozala/unreachable",
+  "repository": "Gozala/unreachable",
   "license": "MIT",
-  "devDependencies": {
-    "babel-cli": "6.24.1",
-    "babel-preset-flow-node": "^1.0.2",
-    "babel-register": "6.24.1",
-    "blue-tape": "^1.0.0",
-    "documentation": "^5.1.0",
-    "flow-bin": "^0.52.0",
-    "flow-copy-source": "^1.2.0",
-    "husky": "^0.14.0",
-    "lint-staged": "^4.0.2",
-    "prettier": "^1.5.3"
-  },
+  "author": "Irakli Gozalishvili <rfobic@gmail.com> (http://jeditoolkit.com)",
   "scripts": {
-    "test": "npm run test:flow && npm run test:tape",
-    "test:tape": "blue-tape -r babel-register test/**/*.js",
-    "test:flow": "flow check",
-    "build:clear": "rm -rf lib",
-    "build:types": "flow-copy-source --verbose src lib",
-    "build:node": "babel --out-dir lib src",
-    "build:api": "documentation readme --section=API src/unreachable.js",
-    "build:docs": "documentation build --document-exported src/** -f html --o docs",
-    "build": "npm run build:node && npm run build:types",
-    "prepublish": "npm run build && npm test",
-    "precommit": "lint-staged",
-    "start": "flow-copy-source --watch --verbose src lib & babel --watch --out-dir lib src"
+    "prepublish": "npm test",
+    "test": "standard && node test && tsd"
   },
-  "lint-staged": {
-    "*.js": ["prettier --parser flow --no-semi --write", "git add"]
+  "devDependencies": {
+    "standard": "^14.3.3",
+    "tsd": "^0.11.0"
+  },
+  "engines": {
+    "node": ">=8.10"
   }
 }
