--- src/unreachable.d.ts
@@ -1,5 +1,0 @@
-export const unreachable = (value: never): never => {
-  throw new TypeError(`Internal error: Encountered impossible value: ${value}`)
-}
-
-export default unreachable
