--- src/unreachable.js
@@ -1,8 +1,0 @@
-/* @flow */
-
-export const unreachable = (value: empty): empty => {
-  const format = JSON.stringify(value)
-  throw new TypeError(`Internal error: Encountered impossible value: ${format}`)
-}
-
-export default unreachable
